Fix concurrency bug with writes to joint states
