+++ components/communication_stack/socket_manager/protobuf_msgs/messages.pb.c
@@ -36,4 +36,3 @@ PB_BIND(OutgoingData, OutgoingData, 2)
 
 
 
-
