+++ components/hardware_drivers/drive_base/drive_base_driver.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "freertos/idf_additions.h"
 #include "hardware_driver.h"
 #include "motor_driver.h"
 
@@ -15,4 +16,12 @@ class DriveBaseDriver : public HardwareDriver
     Motor right_motor_;
 
     bool is_enabled_;
+
+    QueueHandle_t joint_state_publish_queue_;
+    static void publish_timer_callback_(void *arg);
+    esp_timer_handle_t publish_timer_;
+
+    QueueHandle_t joint_cmd_recv_queue_;
+
+    static void task_main_(void *arg);
 };
