+++ components/hardware_drivers/drive_base/motor_driver.h
@@ -2,9 +2,7 @@
 
 #include "driver/pulse_cnt.h"
 #include "esp_timer.h"
-#include "freertos/idf_additions.h"
 #include "hal/ledc_types.h"
-#include "messages.pb.h"
 #include "pid_ctrl.h"
 #include "soc/gpio_num.h"
 
@@ -25,13 +23,13 @@ class Encoder
     float position_;
 
     friend class Motor;
+    friend class DriveBaseDriver;
 };
 
 class Motor
 {
   public:
-    Motor(Joint joint_name,
-          gpio_num_t pwm_a,
+    Motor(gpio_num_t pwm_a,
           ledc_channel_t chan_a,
           gpio_num_t pwm_b,
           ledc_channel_t chan_b,
@@ -42,8 +40,6 @@ class Motor
     void set_velocity(float speed);
 
   private:
-    Joint joint_name_;
-
     Encoder encoder_;
 
     void set_effort_(float power);
@@ -56,7 +52,6 @@ class Motor
     ledc_channel_t chan_b_;
 
     bool reversed_;
-    gpio_num_t enable_pin_;
 
     float cmd_velocity_; // rad / s
     float cmd_position_; // rad
@@ -64,11 +59,5 @@ class Motor
 
     float applied_effort_;
 
-    QueueHandle_t joint_state_publish_queue_;
-    static void publish_timer_callback_(void *arg);
-    esp_timer_handle_t publish_timer_;
-
-    QueueHandle_t joint_cmd_recv_queue_;
-
-    static void task_main_(void *arg);
+    friend class DriveBaseDriver;
 };
